+++ packages/effect/src/interfaces/Pipeable.ts
@@ -550,3 +550,12 @@ export const pipeArguments = <A>(self: A, args: IArguments): unknown => {
     }
   }
 }
+
+/**
+ * @since 4.0.0
+ */
+export const Class: new() => Pipeable = class {
+  pipe() {
+    return pipeArguments(this, arguments)
+  }
+}
