+++ packages/effect/src/schema/ToFormat.ts
@@ -4,7 +4,6 @@
 import type { Format } from "../data/Format.ts"
 import * as Option from "../data/Option.ts"
 import { format, formatPropertyKey } from "../interfaces/Inspectable.ts"
-import { memoizeThunk } from "../internal/schema.ts"
 import type * as Annotations from "./Annotations.ts"
 import * as AST from "./AST.ts"
 import type * as Schema from "./Schema.ts"
@@ -183,7 +182,7 @@ export const defaultReducerAlg: AST.ReducerAlg<Format<any>> = {
     return format(t)
   },
   Suspend: (ast, reduce) => {
-    const get = memoizeThunk(() => reduce(ast.thunk()))
+    const get = AST.memoizeThunk(() => reduce(ast.thunk()))
     return (t) => get()(t)
   }
 }
